cleanunp + fixes
